First stable version version LAMP stack container
In order for containers to be able to communicate between each other, containers need to be linked.  Docker syntax
is important when defining linkage between containers. Links should be defined so that they point to the name of the
container or service they are trying to reach.  Next step is to store environmental variables on a separate file to keep
things secure on production environments.